+++ game/scripts/autoload/minigame_manager.gd
@@ -1,5 +1,7 @@
 extends Node
 
+@export var speed_up_num_rounds: int = 2
+
 @export var levels: Dictionary[String, PackedScene]
 
 @export var tracks: Dictionary[String, MusicTrack]
@@ -13,9 +15,12 @@ var lives_left: int
 var level: int = 1:
 	set(value):
 		level = value
-		level_changed.emit(value)
+		if value > 1:
+			level_changed.emit(value)
 
 var current_level: MinigameLevel
+var last_game_won: bool
+var games_won: int
 
 var previous_levels: Array[String]
 
@@ -26,6 +31,12 @@ var level_duration: float:
 var time_left_in_level: float:
 	get:
 		return level_timer.time_left
+
+func reset() -> void:
+	games_won = 0
+	level = 1
+	last_game_won = false
+	lives_left = 4
 
 func load_random_level() -> void:
 	var level_index: int = randi() % len(levels)
@@ -40,8 +51,6 @@ func load_random_level() -> void:
 	previous_levels.append(level_name)
 	if len(previous_levels) >= len(levels) / 2:
 		previous_levels.remove_at(0)
-	if level < 3:
-		level += 1
 
 func start(time: float = 15) -> void:
 	level_timer.start(time)
@@ -50,20 +59,37 @@ func _on_game_end() -> void:
 	await get_tree().create_timer(4, false).timeout
 	if is_instance_valid(current_level):
 		current_level.game_ended.disconnect(_on_game_end)
+	if games_won > 0 and games_won % speed_up_num_rounds == 0:
+		speed_up()
+	if level > 3:
+		game_win()
+		return
 	if lives_left > 0:
 		load_random_level()
 	else:
 		game_over()
 
+func speed_up() -> void:
+	if level <= 3:
+		level += 1
+
 func end_game() -> void:
 	level_timer.stop()
 	if is_instance_valid(current_level) and not current_level.won_game:
 		current_level.lose()
+	last_game_won = current_level.won_game
+	if last_game_won:
+		games_won += 1
 	current_level.end_game()
 
 func game_over() -> void:
+	print("GAME WIN")
+	current_level = null
 	SceneManager.change_scene(load("uid://vstl3bg568s7"))
-	
+
+func game_win() -> void:
+	print("VICTORY")
+	SceneManager.change_scene(load("uid://vstl3bg568s7"))
 
 func _on_level_timer_timeout() -> void:
 	end_game()