+++ game/scripts/gameplay/ferris_wheel.gd
@@ -20,7 +20,7 @@ var _unhinged: bool
 var _unhinge_spin_speed: float
 
 var _unhinge_velocity_y: float
-var _spinning_player: AudioStreamPlayer2D
+var spinning_player: AudioStreamPlayer2D
 var _roll_player: AudioStreamPlayer2D
 
 func _ready() -> void:
@@ -37,14 +37,15 @@ func _input(event: InputEvent) -> void:
 			_grab(event)
 		elif not event.is_pressed() and is_grabbing:
 			is_grabbing = false
-			if not is_instance_valid(_spinning_player):
-				_spinning_player.stop()
+			if not is_instance_valid(spinning_player):
+				spinning_player.stop()
 	elif is_grabbing and event is InputEventMouseMotion:
 		_spin(event)
 
 func _exit_tree() -> void:
 	if is_instance_valid(_roll_player):
 		_roll_player.queue_free()
+	stop_spin_sound()
 
 func _grab(event: InputEventMouseButton) -> void:
 	is_grabbing = true
@@ -58,10 +59,15 @@ func _spin(event: InputEventMouseMotion) -> void:
 	var grab_offset: Vector2 = normalized_grab_offset * grab_radius
 	var force_vector: Vector2 = event.relative * min(grab_radius, offset.length())
 	body.apply_force(force_vector, grab_offset)
-	if not is_instance_valid(_spinning_player):
-		_spinning_player = AudioManager.play_clip(spin_clip)
+	if not is_instance_valid(spinning_player):
+		spinning_player = AudioManager.play_clip(spin_clip)
 	if not _unhinged and body.angular_velocity > unhinge_angular_velocity:
 		_unhinge()
+
+func stop_spin_sound() -> void:
+	if is_instance_valid(spinning_player):
+		spinning_player.stop()
+		spinning_player.queue_free()
 
 func _unhinge() -> void:
 	grab_collision.disabled = true
@@ -70,9 +76,9 @@ func _unhinge() -> void:
 	pin.queue_free()
 	_unhinged = true
 	unhinge.emit()
-	_spinning_player.queue_free()
+	stop_spin_sound
 	AudioManager.play_clip(unhinge_clip)
-	_roll_player = AudioManager.play_clip(roll_clip)
+	_roll_player = AudioManager.play_clip(roll_clip, 1, 1, 0.9)
 	_unhinge_spin_speed = -body.angular_velocity * 2 * PI * max_grab_radius
 
 func _on_grab_area_area_entered(area: Area2D) -> void: