+++ game/scripts/levels/main_menu/main_menu.gd
@@ -18,11 +18,10 @@ func _ready() -> void:
 	AudioManager.play_music(menu_music, 0, 1)
 	if OS.get_name() == "Web":
 		_quit_button.hide()
-	_menu_buttons.grab_focus()
 
 func _on_start_button_pressed() -> void:
 	#create_tween().tween_property(flowerwall_crt, "modulate", Color.TRANSPARENT, 1).from_current().set_trans(Tween.TRANS_LINEAR)
-	MinigameManager.lives_left = 4
+	MinigameManager.reset()
 	MinigameManager.load_random_level()
 	#SceneManager.change_scene(load("uid://b3neoogxgx3lr"))