+++ game/scripts/levels/spin_the_wheel/spin_the_wheel.gd
@@ -8,6 +8,7 @@ var cursor_scene: PackedScene = preload("uid://bbvasjlghjsv0")
 var scream_clip: AudioStream = preload("uid://ci2ytpvgmqv0q")
 
 func _ready() -> void:
+	ferris_wheel.unhinge_angular_velocity += MinigameManager.level * 3
 	await super._ready()
 	await _start_game(15)
 
@@ -22,10 +23,11 @@ func _start_game(time: float = 10, initial_wait: float = 2) -> void:
 
 func lose() -> void:
 	super.lose()
+	ferris_wheel.stop_spin_sound()
 	ferris_wheel.grab_collision.disabled = true
 
 func _on_ferris_wheel_unhinge() -> void:
-	AudioManager.play_clip(scream_clip)
+	AudioManager.play_clip(scream_clip, 0.85, 1.15, 0.5)
 	win()
 	rotate_dir.hide()
 	MinigameManager.end_game()