+++ game/scripts/ui/pause_menu.gd
@@ -10,7 +10,7 @@ var main_menu_scene: PackedScene = preload("uid://vstl3bg568s7")
 @onready var background_blur: TextureRect = $background_blur
 ## Parent container of menu buttons list
 @onready var margin_container: MarginContainer = $panel/margin_container
-@onready var options_panel: Panel = margin_container.get_node_or_null("options_panel")
+@onready var options_panel: Panel = $options_panel
 ## Parent of the main pause menu buttons
 @onready var menu_buttons: VBoxContainer = margin_container.get_node_or_null("menu_buttons")
 ## A confirmation warning that appears when quitting the game
@@ -71,8 +71,7 @@ func _on_quit_button_pressed() -> void:
 
 func _on_quit_warning_quit_confirmed() -> void:
 	close()
-	MinigameManager.level_timer.stop()
-	SceneManager.change_scene(main_menu_scene)
+	MinigameManager.game_over()
 
 func _on_quit_warning_quit_canceled() -> void:
 	quit_warning.hide()