+++ game/scripts/ui/volume/volume_bar.gd
@@ -6,7 +6,7 @@ func _ready() -> void:
 	MinigameManager.level_changed.connect(_on_level_change)
 
 func set_volume(amount: int) -> void:
-	for i in range(4, amount + 1):
+	for i in range(4, min(10, amount + 1)):
 		await get_tree().create_timer(0.125, false).timeout
 		segments_container.find_child("volume%d" % i).show()